requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
google-generativeai==0.8.3
azure-search-documents==11.6.0b8
azure-storage-blob==12.22.0
pypdf==4.3.1
python-multipart==0.0.12
firebase-admin==6.5.0
google-cloud-firestore==2.19.0
